business-intelligence-bot/requirements.txt
# Core
python-telegram-bot==20.7
python-dotenv==1.0.0

# AI
mistralai==0.3.0
openai==1.12.0

# Database
mysql-connector-python==8.2.0
redis==5.0.1
sqlalchemy==2.0.25
alembic==1.13.1

# Documents
python-docx==1.1.2
reportlab==4.0.9
openpyxl==3.1.2

# Utilities
pandas==2.1.4
numpy==1.26.4
pydantic==2.5.3
python-dateutil==2.8.2
pytz==2023.3.post1

# Web (optional)
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0
